Make use of mirror queue
